refactoring logical_validator, sql_parser to db_entity
