+++ .gitignore
@@ -184,10 +184,6 @@ dbengine-env/
 .gitignore
 
 .vscode/
+src/tests/test_dbapi2.py
+src/server/database/storage/
 src/tests/last_result.csv
-last_result.csv
-src/server/database/storage/CompanyDB/employees.csv
-src/tests/test_dbapi2.py
-src/tests/test_dbapi2.py
-src/server/database/storage/CompanyDB/employees.csv
-src/tests/test_dbapi2.py
