+++ src/server/api/router/query.py
@@ -5,7 +5,8 @@ from fastapi import Depends
 from server.api.schema.query import RequestQuery, ResponseQuery
 from server.middleware.auth import get_current_user
 from server.controllers import db_controlller
-from server.utils.exceptions.http.exc_400 import http_exc_400_query_empty_bad_request
+from server.utils.exceptions import dpapi2_exception
+# from server.utils.exceptions.http.exc_400 import http_exc_400_query_empty_bad_request
 from fastapi.responses import StreamingResponse
 router = fastapi.APIRouter(prefix="/queries", tags=["queries"])
 
@@ -15,8 +16,9 @@ async def query(
     request: RequestQuery,
     current_user = Depends(get_current_user)
 ):
-    if not request.query:
-        raise http_exc_400_query_empty_bad_request()
+    #  check query is in body and not empty
+    if not request.query or not request.query.strip():
+        raise dpapi2_exception.InterfaceError("Empty query provided")
 
     user_query = request.query.lower()
     query_stream = db_controlller.query(
